Configurare roluri Article+Comment

+++ Controllers/ArticlesController.cs
@@ -3,25 +3,51 @@ using ProiectDAW.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Authorization;
 
 namespace ArticlesApp.Controllers
 {
     public class ArticlesController : Controller
-    {
-        private readonly ApplicationDbContext db;
+    {   
+
+        //PAS 10
+
         private readonly IWebHostEnvironment _env;
-        public ArticlesController(ApplicationDbContext context, IWebHostEnvironment env)
+        private readonly ApplicationDbContext db;
+        private readonly UserManager<ApplicationUser> _userManager;
+        private readonly RoleManager<IdentityRole> _roleManager;
+        public ArticlesController(
+        ApplicationDbContext context,
+        UserManager<ApplicationUser> userManager,
+        RoleManager<IdentityRole> roleManager,
+        IWebHostEnvironment env
+        )
         {
-            db = context;
             _env = env;
+            db = context;
+            _userManager = userManager;
+            _roleManager = roleManager;
         }
+
+
+        //private readonly ApplicationDbContext db;
+        //private readonly IWebHostEnvironment _env;
+        //public ArticlesController(ApplicationDbContext context, IWebHostEnvironment env)
+        //{
+        //    db = context;
+        //    _env = env;
+        //}
 
         // Se afiseaza lista tuturor articolelor impreuna cu categoria 
         // din care fac parte
         // HttpGet implicit
+
+        [Authorize(Roles = "Admin,User")]
         public IActionResult Index()
         {
-            var articles = db.Articles.Include("Chapter");
+            var articles = db.Articles.Include("Chapter")
+                                        .Include("User");
 
             // ViewBag.OriceDenumireSugestiva
             ViewBag.Articles = articles;
@@ -29,6 +55,7 @@ namespace ArticlesApp.Controllers
             if (TempData.ContainsKey("message"))
             {
                 ViewBag.Message = TempData["message"];
+                ViewBag.Alert = TempData["messageType"];
             }
 
             return View();
@@ -40,17 +67,62 @@ namespace ArticlesApp.Controllers
         // HttpGet implicit
         public IActionResult Show(int id)
         {
-            Article article = db.Articles.Include("Chapter").Include("Comments")
+            Article article = db.Articles.Include("Chapter")
+                                         .Include("Comments")
+                                         .Include("User")
+                                         .Include("Comments.User")
                               .Where(art => art.Id == id)
                               .First();
 
+            SetAccessRights();
+
             return View(article);
         }
+
+
+
+
+        // Adaugarea unui comentariu asociat unui articol in baza de date
+        // Toate rolurile pot adauga comentarii in baza de date
+        [HttpPost]
+        [Authorize(Roles = "User,Admin")]
+        public IActionResult Show([FromForm] Comment comment)
+        {
+            comment.Date = DateTime.Now;
+
+            // preluam Id-ul utilizatorului care posteaza comentariul
+            comment.UserId = _userManager.GetUserId(User);
+
+            try
+            {
+                db.Comments.Add(comment);
+                db.SaveChanges();
+                return Redirect("/Articles/Show/" + comment.ArticleId);
+            }
+            catch (Exception)
+            {
+                Article art = db.Articles.Include("Chapter")
+                                         .Include("User")
+                                         .Include("Comments")
+                                         .Include("Comments.User")
+                               .Where(art => art.Id == comment.ArticleId)
+                               .First();
+
+                //return Redirect("/Articles/Show/" + comm.ArticleId);
+
+                SetAccessRights();
+
+                return View(art);
+            }
+        }
+
+
 
         // Se afiseaza formularul in care se vor completa datele unui articol
         // impreuna cu selectarea categoriei din care face parte
         // HttpGet implicit
 
+        [Authorize(Roles ="User,Admin")]
         public IActionResult New()
         {
           
@@ -62,10 +134,15 @@ namespace ArticlesApp.Controllers
         }
 
         // POST: Procesează datele trimise de utilizator
+
         [HttpPost]
+        [Authorize(Roles = "User,Admin")]
         public async Task<IActionResult> New(Article article, IFormFile Image)
         {
             article.Date = DateTime.Now;
+
+            //preluam id-ul user-ului
+            article.UserId = _userManager.GetUserId(User);
 
  
             if (Image != null && Image.Length > 0)
@@ -100,6 +177,9 @@ namespace ArticlesApp.Controllers
                 db.Articles.Add(article);
                 await db.SaveChangesAsync();
 
+                TempData["message"] = "Articolul a fost adaugat";
+                TempData["messageType"] = "alert-success";
+
                 // Redirecționare după succes
                 return RedirectToAction("Index", "Articles");
             }
@@ -108,11 +188,13 @@ namespace ArticlesApp.Controllers
             return View(article);
         }
 
-       
+
         // Se editeaza un articol existent in baza de date impreuna cu categoria din care face parte
         // Categoria se selecteaza dintr-un dropdown
         // HttpGet implicit
         // Se afiseaza formularul impreuna cu datele aferente articolului din baza de date
+
+        [Authorize(Roles ="Admin,User")]
         public IActionResult Edit(int id)
         {
 
@@ -122,25 +204,53 @@ namespace ArticlesApp.Controllers
 
             article.Chap = GetAllChapters();
 
-            return View(article);
+            if( (article.UserId== _userManager.GetUserId(User)) || User.IsInRole("Admin"))
+            { 
+                return View(article);
+            }
+            else
+            {
+                
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine";
+                TempData["messageType"] = "alert-danger";
 
+                return RedirectToAction("Index");
+                //return RedirectToAction("Show", new { id });
+            }
         }
 
         // Se adauga articolul modificat in baza de date
         [HttpPost]
+        [Authorize(Roles = "Admin,User")]
         public IActionResult Edit(int id, Article requestArticle)
         {
             Article article = db.Articles.Find(id);
 
             try
             {
-                article.Title = requestArticle.Title;
-                article.Content = requestArticle.Content;
-                article.Date = DateTime.Now;
-                article.ChapterId = requestArticle.ChapterId;
-                db.SaveChanges();
-                TempData["message"] = "Articolul a fost modificat";
-                return RedirectToAction("Index");
+                if ((article.UserId == _userManager.GetUserId(User)) || User.IsInRole("Admin"))
+                {
+
+                    article.Title = requestArticle.Title;
+                    article.Content = requestArticle.Content;
+                    article.Date = DateTime.Now;
+                    article.ChapterId = requestArticle.ChapterId;
+                    db.SaveChanges();
+                    TempData["message"] = "Articolul a fost modificat";
+                    TempData["messageType"] = "alert-success";
+
+                    return RedirectToAction("Index");
+                    //return RedirectToAction("Show", new { id });
+
+                }
+                else
+                {
+                    TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine";
+                    TempData["messageType"] = "alert-danger";
+
+                    return RedirectToAction("Index");
+                    //return RedirectToAction("Show", new { id });
+                }
 
             }
             catch (Exception e)
@@ -153,14 +263,61 @@ namespace ArticlesApp.Controllers
 
         // Se sterge un articol din baza de date 
         [HttpPost]
+        [Authorize(Roles = "Admin,User")]
         public ActionResult Delete(int id)
         {
-            Article article = db.Articles.Find(id);
-            db.Articles.Remove(article);
-            db.SaveChanges();
-            TempData["message"] = "Articolul a fost sters";
-            return RedirectToAction("Index");
+            Article article = db.Articles.Include("Comments")
+                                         .Where(art => art.Id == id)
+                                         .First();
+
+            if ((article.UserId == _userManager.GetUserId(User)) || User.IsInRole("Admin"))
+            {
+
+                // Construiește calea completă către imagine
+                var filePath = Path.Combine(_env.WebRootPath, article.Image.TrimStart('/'));
+
+                // Șterge imaginea dacă există
+                if (System.IO.File.Exists(filePath))
+                {
+                    System.IO.File.Delete(filePath);
+                }
+
+                db.Articles.Remove(article);
+                db.SaveChanges();
+                TempData["message"] = "Articolul a fost sters";
+                TempData["messageType"] = "alert-success";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine";
+                TempData["messageType"] = "alert-danger";
+
+                return RedirectToAction("Index");
+                //return RedirectToAction("Show", new { id });
+            }
         }
+
+
+
+        // Conditiile de afisare pentru butoanele de editare si stergere
+        // butoanele aflate in view-uri
+        private void SetAccessRights()
+        {
+            ViewBag.AfisareButoane = false;
+
+            if (User.IsInRole("User"))
+            {
+                ViewBag.AfisareButoane = true;
+            }
+
+            ViewBag.UserCurent = _userManager.GetUserId(User);
+
+            ViewBag.EsteAdmin = User.IsInRole("Admin");
+        }
+
+
+
 
         [NonAction]
         public IEnumerable<SelectListItem> GetAllChapters()
